% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_method.R
\name{choose_best_method}
\alias{choose_best_method}
\title{Choose Best Method}
\usage{
choose_best_method(n, r_matrix)
}
\arguments{
\item{n}{Integer, sample size.}

\item{r_matrix}{Square correlation matrix.}
}
\value{
A list with the best method name and details.
}
\description{
Selects the best correlation comparison method based on sample size and average inter-correlation.
}
\examples{
r <- matrix(c(1, 0.40, 0.35,
              0.40, 1, 0.30,
              0.35, 0.30, 1), nrow = 3)
choose_best_method(n = 250, r_matrix = r)
}
