% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scale_network_analysis.R
\name{print.scale_network_analysis}
\alias{print.scale_network_analysis}
\title{Print method for scale_network_analysis objects}
\usage{
\method{print}{scale_network_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{scale_network_analysis}.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for scale_network_analysis objects
}
