% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centrality_sd_correlation.R
\name{calculate_centrality_sd_correlation}
\alias{calculate_centrality_sd_correlation}
\title{Centrality-SD Correlation}
\usage{
calculate_centrality_sd_correlation(Data, Centralitys)
}
\arguments{
\item{Data}{Data frame with the original data.}

\item{Centralitys}{Centrality results object with a \code{table} element.}
}
\description{
Calculates correlation between centrality indices and standard deviations.
}
