% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_EGA_list_to_df.R
\name{convert_EGA_list_to_df}
\alias{convert_EGA_list_to_df}
\title{Convert EGA List to Data Frame}
\usage{
convert_EGA_list_to_df(ega_list)
}
\arguments{
\item{ega_list}{Named list of EGA result objects.}
}
\description{
Converts a list of EGA results into a combined data frame.
}
