% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_EGA_to_df.R
\name{convert_EGA_to_df}
\alias{convert_EGA_to_df}
\title{Convert EGA Result to Data Frame}
\usage{
convert_EGA_to_df(ega_result)
}
\arguments{
\item{ega_result}{A single EGA result object.}
}
\description{
Extracts network metrics from a single EGA result object.
}
