% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_report.R
\name{Density_report}
\alias{Density_report}
\title{Network Density Report}
\usage{
Density_report(edge.matrix)
}
\arguments{
\item{edge.matrix}{Network edge matrix (e.g., from qgraph).}
}
\description{
Calculates and reports network density statistics.
}
