% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_correlation_stability.R
\name{filter_correlation_stability}
\alias{filter_correlation_stability}
\title{Filter Correlation Stability}
\usage{
filter_correlation_stability(caseDroppingBoot)
}
\arguments{
\item{caseDroppingBoot}{Bootstrap results from bootnet case-dropping.}
}
\description{
Extracts and filters correlation stability indices from bootstrap results.
}
