% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_weights_summary.R
\name{get_edge_weights_summary}
\alias{get_edge_weights_summary}
\title{Edge Weights Summary}
\usage{
get_edge_weights_summary(network, abs_weights = T, round_digits = 2)
}
\arguments{
\item{network}{Network object with a graph element.}

\item{abs_weights}{Logical; use absolute edge weights.}

\item{round_digits}{Number of decimal places for rounding.}
}
\description{
Computes descriptive statistics for network edge weights.
}
