% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgm_error_metrics.R
\name{mgm_error_metrics}
\alias{mgm_error_metrics}
\title{MGM Error Metrics}
\usage{
mgm_error_metrics(data, type, level, k = 2)
}
\arguments{
\item{data}{Data frame or matrix with the data.}

\item{type}{Character vector of variable types ("g", "c", "p").}

\item{level}{Integer vector of variable levels.}

\item{k}{Order of interactions (default: 2).}
}
\description{
Fits a Mixed Graphical Model and extracts prediction error metrics.
}
