% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgm_errors_groups.R
\name{mgm_errors_groups}
\alias{mgm_errors_groups}
\title{MGM Errors by Group}
\usage{
mgm_errors_groups(data, type, level, group, columns)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{type}{Character vector of variable types.}

\item{level}{Integer vector of variable levels.}

\item{group}{Unquoted grouping variable name.}

\item{columns}{Character vector of column names to analyze.}
}
\description{
Fits Mixed Graphical Models and extracts errors for each group.
}
