% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_reduce2.R
\name{net_reduce2}
\alias{net_reduce2}
\title{Network Reduction}
\usage{
net_reduce2(
  data,
  badpairs,
  method = c("PCA", "best_goldbricker"),
  keep_unpaired = TRUE
)
}
\arguments{
\item{data}{Data frame with item responses.}

\item{badpairs}{Goldbricker object or list of redundant node pairs.}

\item{method}{Reduction method: "PCA" or "best_goldbricker".}

\item{keep_unpaired}{Logical; keep unpaired items in result.}
}
\description{
Reduces redundant node pairs using PCA or best goldbricker selection.
}
