% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_EGA_combinations.R
\name{run_EGA_combinations}
\alias{run_EGA_combinations}
\title{Run EGA Combinations}
\usage{
run_EGA_combinations(
  data,
  corr = c("cor_auto", "pearson", "spearman"),
  algorithm = c("leiden", "louvain", "walktrap"),
  leiden_args = list(objective_function = "CPM", resolution_parameter = 1)
)
}
\arguments{
\item{data}{Data frame with the data.}

\item{corr}{Character vector of correlation methods.}

\item{algorithm}{Character vector of community detection algorithms.}

\item{leiden_args}{List of additional arguments for the Leiden algorithm.}
}
\value{
A list of EGA results with class "EGA_combinations".
}
\description{
Runs EGA with multiple combinations of correlation methods and algorithms.
}
