% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{summary_metrics}
\alias{summary_metrics}
\title{Summarize Network Performance Metrics}
\usage{
summary_metrics(results)
}
\arguments{
\item{results}{Data frame from boot_and_evaluate with raw metrics.}
}
\value{
A data frame with averaged metrics.
}
\description{
Calculates median values of network performance metrics
grouped by algorithm, correlation method, and sample size.
}
